modify and add ICMPv6

+++ layer3.h
@@ -1,7 +1,7 @@
 #ifndef __LAYER3_H
 #define __LAYER3_H
 
-u_int8_t layer3(u_int);
+int layer3(u_int);
 u_int8_t disp_ipv4(void);
 u_int8_t disp_ipv6(void);
